mood_tracker_agent/pyproject.toml
[project]
name = "mood-tracker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.0",
    "python-dotenv>=1.1.1",
]
